refactor: make wasm compile
